profiles/models/profile_photo/serializers.py
from rest_framework import serializers

from profiles.models import ProfilePhoto


class ProfilePhotoSerializer(serializers.ModelSerializer):
    class Meta:
        model = ProfilePhoto
        fields = '__all__'
