requirements.txt
asgiref==3.3.4
backcall==0.2.0
decorator==5.0.9
Django==3.2.5
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-settings-export==1.2.1
django-split-settings==1.0.1
djangorestframework==3.12.4
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkg_resources==0.0.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
